+++ src/main/java/com/tristankechlo/livingthings/config/LivingThingsConfig.java
@@ -18,6 +18,7 @@ import com.tristankechlo.livingthings.config.entity.OwlConfig;
 import com.tristankechlo.livingthings.config.entity.PenguinConfig;
 import com.tristankechlo.livingthings.config.entity.RaccoonConfig;
 import com.tristankechlo.livingthings.config.entity.SharkConfig;
+import com.tristankechlo.livingthings.config.entity.ShroomieConfig;
 import com.tristankechlo.livingthings.config.entity.SnailConfig;
 
 import net.minecraft.util.ResourceLocation;
@@ -63,6 +64,7 @@ public class LivingThingsConfig {
 	public static final SnailConfig SNAIL = new SnailConfig(BUILDER);
 	public static final MonkeyConfig MONKEY = new MonkeyConfig(BUILDER);
 	public static final NetherKnightConfig NETHER_KNIGHT = new NetherKnightConfig(BUILDER);
+	public static final ShroomieConfig SHROOMIE = new ShroomieConfig(BUILDER);
 
 	public static final ForgeConfigSpec spec = BUILDER.build();