+++ src/main/java/com/tristankechlo/livingthings/events/RegistryEventHandler.java
@@ -16,12 +16,13 @@ import com.tristankechlo.livingthings.entities.OwlEntity;
 import com.tristankechlo.livingthings.entities.PenguinEntity;
 import com.tristankechlo.livingthings.entities.RaccoonEntity;
 import com.tristankechlo.livingthings.entities.SharkEntity;
+import com.tristankechlo.livingthings.entities.ShroomieEntity;
 import com.tristankechlo.livingthings.entities.SnailEntity;
 import com.tristankechlo.livingthings.init.ModEntityTypes;
 
 import net.minecraft.entity.EntitySpawnPlacementRegistry;
-import net.minecraft.entity.EntityType;
 import net.minecraft.entity.EntitySpawnPlacementRegistry.PlacementType;
+import net.minecraft.entity.EntityType;
 import net.minecraft.entity.monster.MonsterEntity;
 import net.minecraft.entity.passive.AnimalEntity;
 import net.minecraft.world.gen.Heightmap;
@@ -57,6 +58,7 @@ public class RegistryEventHandler {
 		event.put(ModEntityTypes.SNAIL_ENTITY.get(), SnailEntity.createAttributes().build());
 		event.put(ModEntityTypes.MONKEY_ENTITY.get(), MonkeyEntity.createAttributes().build());
 		event.put(ModEntityTypes.NETHER_KNIGHT_ENTITY.get(), NetherKnightEntity.createAttributes().build());
+		event.put(ModEntityTypes.SHROOMIE_ENTITY.get(), ShroomieEntity.createAttributes().build());
 	}
 
 	private static void registerEntitySpawnPlacements() {
@@ -76,6 +78,7 @@ public class RegistryEventHandler {
 		EntitySpawnPlacementRegistry.register(ModEntityTypes.SNAIL_ENTITY.get(), PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, AnimalEntity::checkAnimalSpawnRules);
 		EntitySpawnPlacementRegistry.register(ModEntityTypes.MONKEY_ENTITY.get(), PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING, MonkeyEntity::canMonkeySpawn);
 		EntitySpawnPlacementRegistry.register(ModEntityTypes.NETHER_KNIGHT_ENTITY.get(), PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, MonsterEntity::checkMonsterSpawnRules);
+		EntitySpawnPlacementRegistry.register(ModEntityTypes.SHROOMIE_ENTITY.get(), PlacementType.ON_GROUND, Heightmap.Type.MOTION_BLOCKING_NO_LEAVES, ShroomieEntity::canShroomieSpawn);
 	}
 
 }