+++ Changelog.md
@@ -3,6 +3,7 @@
 ### Version 1.16.5 - 1.3.3
  - compress images
  - compress sound files
+ - add shroomie-entity
 
 ### Version 1.16.5 - 1.3.2
  - add nether knight entity